Small fix in NDimTable.cpp

+++ lib/base/NDimTable.cpp
@@ -4,7 +4,7 @@
 
 std::ostream & operator<<(std::ostream &os, const std::vector<std::size_t>& dim)
 {
-	for(int i=0 ; i<dim.size() ; i++ )
+	for(size_t i=0 ; i<dim.size() ; i++ )
 	{
 		os << dim[i] ;
 		if(i!=dim.size()-1) os << ",";
